Renamed file to grammar-reference-finder.ts
