+++ packages/langium/src/grammar/lsp/grammar-reference-finder.ts
